chore(quality): adopt the busirocket Python baseline in qlctool
The whole gate applies and exits 0. What it took, beyond the mechanical
formatting pass already committed:

- Four defects it found, all fixed: speed_dial annotated
  Sequence["DialFunction"] without importing the name, so anything that
  resolves annotations broke; movement_efx shadowed its fixture_ids
  parameter with the per-part loop variable; one unused loop binding and
  one unused local in a test. Three dead imports removed, and
  test_stage_layout now reads POINTS_OF_VIEW from monitor_node, where it
  lives, instead of through stage_layout's accidental re-export.
- lxml-stubs installed rather than silenced: it cleared 67 of the 303
  mypy errors outright. The remaining 266 are a dated per-module
  ignore_errors ratchet across 73 modules.
- 170 structural findings recorded as baseline debt, with one reasoned
  override: CLAUDE.md defines the check-rule convention (a rule lives in
  checks/rule_<name>.py and exposes check_<name>), so BPY002 is off for
  those eight files and on everywhere else.
- Coverage measured 86.51%; the floor is 85, above the shared 80, so the
  gain cannot drain away unnoticed.
- The ruff ratchet in ruff.toml is dated and per-family, with T201 and
  PLC0415 as permanent decisions: qlctool is a command that reports with
  print and defers its heavy imports.

The venv was rebuilt without --system-site-packages, which is why
pip-audit now passes: it had been auditing every Homebrew package on
this Mac, and the tests could have been passing on a package absent from
the lockfile. They pass without it - 352 tests - and qlctool check on
Vibra-split.qxw still reports 503 buttons with no problems.

+++ qlctool/generate/beat_tempo.py
@@ -91,4 +91,4 @@ def _to_beats(function, timing: BeatTiming) -> None:
 
 def _units(beats: float) -> int:
     """Beats to QLC+'s thousandths, on its own eighth-of-a-beat grid."""
-    return int(round(beats * UNITS_PER_BEAT / QUANTUM)) * QUANTUM
+    return round(beats * UNITS_PER_BEAT / QUANTUM) * QUANTUM
